# docs/swagger.yaml
definitions:
...
  handlers.CreateContainerRequest:
    properties:
      image:
        type: string
      snapshotter:
        type: string
    type: object
  handlers.CreateContainerResponse:
    properties:
      container_id:
        type: string
      image:
        type: string
      snapshotter:
        type: string
      started:
        type: boolean
    type: object
  handlers.CreateSnapshotRequest:
    properties:
      snapshot_name:
        type: string
    type: object
  handlers.CreateSnapshotResponse:
    properties:
      container_id:
        type: string
      snapshot_name:
        type: string
      snapshotter:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handlers.GetContainerResponse:
    properties:
      container_id:
        type: string
      container_path:
        type: string
      created_at:
        type: string
      host_path:
        type: string
      image:
        type: string
      namespace:
        type: string
      status:
        type: string
      task_running:
        type: boolean
      updated_at:
        type: string
    type: object
  handlers.ListSnapshotsResponse:
    properties:
      snapshots:
        items:
          $ref: '#/definitions/handlers.SnapshotInfo'
        type: array
      snapshotter:
        type: string
    type: object
...
  handlers.SkillItem:
    properties:
      content:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  handlers.SkillsDeleteRequest:
    properties:
      names:
        items:
          type: string
        type: array
    type: object
  handlers.SkillsResponse:
    properties:
      skills:
        items:
          $ref: '#/definitions/handlers.SkillItem'
        type: array
    type: object
  handlers.SkillsUpsertRequest:
    properties:
      skills:
        items:
          $ref: '#/definitions/handlers.SkillItem'
        type: array
    type: object
  handlers.SnapshotInfo:
    properties:
      created_at:
        type: string
      kind:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      parent:
        type: string
      snapshotter:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.skillsOpResponse:
    properties:
      ok:
        type: boolean
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/container:
    delete:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete MCP container for bot
      tags:
      - containerd
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetContainerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get container info for bot
      tags:
      - containerd
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Create container payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateContainerRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CreateContainerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create and start MCP container for bot
      tags:
      - containerd
  /bots/{bot_id}/container/snapshots:
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Snapshotter name
        in: query
        name: snapshotter
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListSnapshotsResponse'
      summary: List snapshots
      tags:
      - containerd
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Create snapshot payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSnapshotRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CreateSnapshotResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create container snapshot for bot
      tags:
      - containerd
  /bots/{bot_id}/container/start:
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Start container task for bot
      tags:
      - containerd
  /bots/{bot_id}/container/stop:
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Stop container task for bot
      tags:
      - containerd
...
  /container/skills:
    delete:
      parameters:
      - description: Delete skills payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.SkillsDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.skillsOpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete skills from container
      tags:
      - containerd
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SkillsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List skills from container
      tags:
      - containerd
    post:
      parameters:
      - description: Skills payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.SkillsUpsertRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.skillsOpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Upload skills into container
      tags:
      - containerd
...
swagger: "2.0"